requirements.txt
torch
torchvision
Pillow
streamlit
requests
python-dotenv
rembq
albumentations